requirements.txt
opencv-python-headless
pytesseract
numpy

Pillow
